clean variants for benchmarks
