+++ src/benchmarks/limbo/bench.cpp
@@ -57,7 +57,7 @@ struct Params {
         BO_PARAM(bool, stats_enabled, false);
     };
     struct bayes_opt_boptimizer {
-        BO_PARAM(int, hp_period, -1);
+        BO_PARAM(int, hp_period, 50);
     };
     struct stop_maxiterations {
         BO_PARAM(int, iterations, 190);
@@ -65,18 +65,30 @@ struct Params {
     struct kernel : public defaults::kernel {
         BO_PARAM(double, noise, 1e-10);
     };
+    struct kernel_exp : public defaults::kernel_exp {
+    };
     struct kernel_maternfivehalves {
         BO_PARAM(double, sigma_sq, 1);
         BO_PARAM(double, l, 1);
     };
-    struct acqui_ucb {
+    struct acqui_ucb : public defaults::acqui_ucb {
         BO_PARAM(double, alpha, 0.125);
+    };
+    struct acqui_ei : public defaults::acqui_ei {
     };
     struct init_randomsampling {
         BO_PARAM(int, samples, 10);
     };
+    struct kernel_squared_exp_ard : public defaults::kernel_squared_exp_ard {
+    };
     struct mean_constant {
         BO_PARAM(double, constant, 1);
+    };
+    struct opt_rprop : public defaults::opt_rprop {
+    };
+    struct opt_parallelrepeater : public defaults::opt_parallelrepeater {
+    };
+    struct opt_nloptnograd : public defaults::opt_nloptnograd {
     };
 };
 
@@ -91,9 +103,15 @@ struct BobyqaParams {
         BO_DYN_PARAM(int, iterations);
     };
 };
+struct BobyqaParams_HP {
+    struct opt_nloptnograd : public defaults::opt_nloptnograd {
+        BO_DYN_PARAM(int, iterations);
+    };
+};
 
 BO_DECLARE_DYN_PARAM(int, DirectParams::opt_nloptnograd, iterations);
 BO_DECLARE_DYN_PARAM(int, BobyqaParams::opt_nloptnograd, iterations);
+BO_DECLARE_DYN_PARAM(int, BobyqaParams_HP::opt_nloptnograd, iterations);
 
 template <typename Optimizer, typename Function>
 void benchmark(const std::string& name)
@@ -101,6 +119,8 @@ void benchmark(const std::string& name)
     int iters_base = 250;
     DirectParams::opt_nloptnograd::set_iterations(static_cast<int>(iters_base * Function::dim_in() * 0.9));
     BobyqaParams::opt_nloptnograd::set_iterations(iters_base * Function::dim_in() - DirectParams::opt_nloptnograd::iterations());
+
+    BobyqaParams_HP::opt_nloptnograd::set_iterations(10 * Function::dim_in() * Function::dim_in());
 
     auto t1 = std::chrono::steady_clock::now();
     Optimizer opt;
@@ -127,6 +147,26 @@ int main()
 {
     srand(time(NULL));
 
+
+// limbo default parameters
+#ifdef LIMBO_DEF
+    using Opt_t = bayes_opt::BOptimizer<Params>;
+#elif defined(LIMBO_DEF_HPOPT)
+    using Opt_t = bayes_opt::BOptimizerHPOpt<Params>;
+
+// Bayesopt default parameters
+#elif defined(BAYESOPT_DEF_HPOPT)
+    using Kernel_t = kernel::SquaredExpARD<Params>;
+    using AcquiOpt_t = opt::Chained<Params, opt::NLOptNoGrad<DirectParams, nlopt::GN_DIRECT_L>, opt::NLOptNoGrad<BobyqaParams, nlopt::LN_BOBYQA>>;
+    using Stop_t = boost::fusion::vector<stop::MaxIterations<Params>>;
+    using Mean_t = mean::Constant<Params>;
+    using Stat_t = boost::fusion::vector<>;
+    using Init_t = init::RandomSampling<Params>;
+    using GP_t = model::GP<Params, Kernel_t, Mean_t, model::gp::KernelLFOpt<Params, opt::NLOptNoGrad<BobyqaParams_HP, nlopt::LN_BOBYQA>>>;
+    using Acqui_t = acqui::UCB<Params, GP_t>;
+
+    using Opt_t = bayes_opt::BOptimizer<Params, modelfun<GP_t>, initfun<Init_t>, acquifun<Acqui_t>, acquiopt<AcquiOpt_t>, statsfun<Stat_t>, stopcrit<Stop_t>>;
+#elif defined(BAYESOPT_DEF)
     using Kernel_t = kernel::MaternFiveHalves<Params>;
     using AcquiOpt_t = opt::Chained<Params, opt::NLOptNoGrad<DirectParams, nlopt::GN_DIRECT_L>, opt::NLOptNoGrad<BobyqaParams, nlopt::LN_BOBYQA>>;
     using Stop_t = boost::fusion::vector<stop::MaxIterations<Params>>;
@@ -137,6 +177,27 @@ int main()
     using Acqui_t = acqui::UCB<Params, GP_t>;
 
     using Opt_t = bayes_opt::BOptimizer<Params, modelfun<GP_t>, initfun<Init_t>, acquifun<Acqui_t>, acquiopt<AcquiOpt_t>, statsfun<Stat_t>, stopcrit<Stop_t>>;
+
+// benchmark different optimization algorithms
+#elif defined(OPT_CMAES)
+    using AcquiOpt_t = opt::Chained<Params, opt::NLOptNoGrad<DirectParams, nlopt::GN_DIRECT_L>, opt::NLOptNoGrad<BobyqaParams, nlopt::LN_BOBYQA>>;
+    using Opt_t = bayes_opt::BOptimizer<Params, acquiopt<AcquiOpt_t>>;
+#elif defined(OPT_DIRECT)
+    using AcquiOpt_t = opt::Chained<Params, opt::NLOptNoGrad<DirectParams, nlopt::GN_DIRECT_L>, opt::NLOptNoGrad<BobyqaParams, nlopt::LN_BOBYQA>>;
+    using Opt_t = bayes_opt::BOptimizer<Params, acquiopt<AcquiOpt_t>>;
+
+//benchmark different acquisition functions
+#elif defined(ACQ_UCB)
+    using GP_t = model::GP<Params>;
+    using Acqui_t = acqui::UCB<Params, GP_t>;
+    using Opt_t =  bayes_opt::BOptimizer<Params, acquifun<Acqui_t>> ;
+#elif defined(ACQ_EI)
+    using GP_t = model::GP<Params>;
+    using Acqui_t = acqui::EI<Params, GP_t>;
+    using Opt_t =  bayes_opt::BOptimizer<Params, acquifun<Acqui_t>> ;
+#else
+    #error "Unknown variant in benchmark"
+#endif
 
     benchmark<Opt_t, BraninNormalized>("branin");
     benchmark<Opt_t, Hartmann6>("hartmann6");
